# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: KlassenPilot API
  version: 0.1.0
paths:
...
  /api/classes/{class_id}/course/materials:
    get:
      tags:
      - course-materials
      summary: Materials
      operationId: materials_api_classes__class_id__course_materials_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/material-imports:
    get:
      tags:
      - course-materials
      summary: Imports
      operationId: imports_api_classes__class_id__course_material_imports_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags:
      - course-materials
      summary: Upload
      operationId: upload_api_classes__class_id__course_material_imports_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_api_classes__class_id__course_material_imports_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/material-imports/{draft_id}:
    get:
      tags:
      - course-materials
      summary: Import Status
      operationId: import_status_api_classes__class_id__course_material_imports__draft_id__get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags:
      - course-materials
      summary: Update Import
      operationId: update_import_api_classes__class_id__course_material_imports__draft_id__put
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialEdit'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/material-imports/{draft_id}/retry:
    post:
      tags:
      - course-materials
      summary: Retry Import
      operationId: retry_import_api_classes__class_id__course_material_imports__draft_id__retry_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/material-imports/{draft_id}/review:
    post:
      tags:
      - course-materials
      summary: Review Import
      operationId: review_import_api_classes__class_id__course_material_imports__draft_id__review_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/material-imports/{draft_id}/approve:
    post:
      tags:
      - course-materials
      summary: Approve Import
      operationId: approve_import_api_classes__class_id__course_material_imports__draft_id__approve_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: draft_id
        in: path
        required: true
        schema:
          type: string
          title: Draft Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Snapshot'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/materials/{material_id}/sections/{section_id}:
    get:
      tags:
      - course-materials
      summary: Section
      operationId: section_api_classes__class_id__course_materials__material_id__sections__section_id__get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: material_id
        in: path
        required: true
        schema:
          type: string
          title: Material Id
      - name: section_id
        in: path
        required: true
        schema:
          type: string
          title: Section Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/materials/{material_id}/files/{asset}:
    get:
      tags:
      - course-materials
      summary: Asset
      operationId: asset_api_classes__class_id__course_materials__material_id__files__asset__get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: material_id
        in: path
        required: true
        schema:
          type: string
          title: Material Id
      - name: asset
        in: path
        required: true
        schema:
          type: string
          title: Asset
      - name: original_page
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Original Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/changes/generate:
    post:
      tags:
      - course-materials
      summary: Generate Changes
      operationId: generate_changes_api_classes__class_id__course_changes_generate_post
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseGenerationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Unusable or timed-out model response; retry without changing memory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/classes/{class_id}/course/changes:
    get:
      tags:
      - course-materials
      summary: List Changes
      operationId: list_changes_api_classes__class_id__course_changes_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Class, draft, or approved material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale review, inactive draft, or publication conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid source, page selection, extraction, or graph change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorEnvelope
...
    Snapshot:
      properties:
        expected_revision:
          type: integer
          title: Expected Revision
        expected_hash:
          type: string
          title: Expected Hash
      type: object
      required:
      - expected_revision
      - expected_hash
      title: Snapshot
...
    CourseGenerationRequest:
      properties:
        purpose:
          type: string
          enum:
          - curriculum_draft
          - material_enrichment
          - correction
          title: Purpose
        teacher_request:
          type: string
          maxLength: 6000
          title: Teacher Request
          default: ''
        material_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Material Id
      additionalProperties: false
      type: object
      required:
      - purpose
      title: CourseGenerationRequest
    MaterialEdit:
      properties:
        expected_revision:
          type: integer
          title: Expected Revision
        expected_hash:
          type: string
          title: Expected Hash
        artifact:
          $ref: '#/components/schemas/MaterialImportArtifact'
      type: object
      required:
      - expected_revision
      - expected_hash
      - artifact
      title: MaterialEdit
...
    Body_upload_api_classes__class_id__course_material_imports_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        title:
          type: string
          title: Title
          default: ''
        arm:
          type: string
          title: Arm
          default: personal
        pages:
          type: string
          title: Pages
          default: ''
      type: object
      required:
      - file
      title: Body_upload_api_classes__class_id__course_material_imports_post
...